+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support for `questionpart.gapfill.gap.choose_one` default file
 - Support for `jsx_graph` extension
+- Support for `match_answers` part type
 ### Changed
 - Type of marks in multiple choice answer (from float to primitive)
 - Type of parameters in function (from HashMap to Vec to preserve order)